data_pipeline/airflow/requirements.txt
apache-airflow-providers-google
pyarrow
pandas
numpy
sqlalchemy
kaggle
psycopg2-binary